Write all objects to ModuleDataStore as atomic
This also removes deleteInvalidModuleData, as it is an unnecesasry
operation for the cache -- in the case where there is invalid data, we
would already be returning an error and the contents of the cache will
be flushed and not reused.

This operation is still needed for the cache using tar files, so we
moved that logic up to getModuleDataForModuleKey.

This is likely also a cause for cache corruption, since it causes a
racey situation.